maybe fix hildevice double processing of nv responses
this issue was exposed when i added a test where i called the module_to_nv()
method to trigger an nv save.  in this method the save parameter is written to
0, then 1, then 0.  the symptom was that it instead wrote 0, then 0, then
raised an exception because it thought it had written 1 and the embedded
responded 0.  basically, it was receiving the response in two different ways
(twisted protocol vs. classic nv objects).  whether there was an issue depended
on the order in which those two receipt paths were executed.  in the bad
scenario...  the test wrote 0, received the response and processed one path
that allowed the test to move forward, set the nv object to 1, processed the
second response path which set the nv object back to 0, sent 0 while thinking
it was sending one, failed when the inverter (correctly) responded 0.

maybe this fix is ok if we just ban reading values from the nv directly which
we probably should for a clean interface definition in the hildevice.  but...
i dunno.