chore(SC-2522): Update pyelftools to be compatible with dgt eabi parser
To be able to include the parse_eabi_file.py from dgt, the pyelftools has to be
updated to same version. Since epyq symbol parsing does not work anyway, the
cmemoryparser is fixed just enough that it does not cause interpreter errors
when launching it. The cmemoryparser or epyq datalogger will not be used in
the foreseeable future in M